requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
beautifulsoup4
selenium
webdriver-manager
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dotenv